+++ sdk/cosmos/azure_data_cosmos/src/clients/container_client.rs
@@ -681,7 +681,7 @@ impl ContainerClient {
     /// ```
     ///
     /// See [`PartitionKey`](crate::PartitionKey) for more information on how to specify a partition key, and [`Query`] for more information on how to specify a query.
-    pub fn query_items<T: DeserializeOwned + 'static>(
+    pub fn query_items<T: DeserializeOwned + Send + 'static>(
         &self,
         query: impl Into<Query>,
         partition_key: impl Into<PartitionKey>,