+++ sdk/keyvault/azure_security_keyvault_secrets/CHANGELOG.md
@@ -8,7 +8,13 @@
 
 - Changed model fields defined as an `Option<HashMap<K, V>>` to just a `HashMap<K, V>`.
 - Changed model fields defined as an `Option<Vec<T>>` to just a `Vec<T>`.
-- Renamed `SecretClient::get_secrets` to `list_secrets`.
+- Renamed `DeletedSecretBundle` to `DeleteSecret`.
+- Renamed `DeletedSecretItem` to `DeletedSecretProperties`.
+- Renamed `SecretBundle` to `Secret`.
+- Renamed `SecretBundleBackup` to `SecretBackup`.
+- Renamed `SecretClient::get_secrets` to `list_secret_properties`.
+- Renamed `SecretItem` to `SecretProperties`.
+- Renamed all parameter types to match the *{Verb}{Noun}* format of the client methods that use them e.g., `set_secret` uses `SetSecretParameters`.
 
 ### Bugs Fixed